Add skipped tests for wishful behavior
That the @ee.on(...) decorator does the "right" thing when it comes to methods on classes.
 - @staticmethod - should work out-of-the-box due to how the decorator works
 - @classmethod - preferably a _bound_ classmethod listener is added
 - instance methods - preferably the _bound_ instance methods are only added when the class is instatiated

I don't know if this behavior would be a good idea as it'd would break the expectations of how decorators normally work, and from experiments be rather complex. Maybe a keyword argument to opt-in to the behavior?